requirements.txt
flask
openai>=1.0.0
chromadb
python-dotenv
pandas
openpyxl
tqdm
